+++ lib/authlogic/acts_as_authentic/single_access_token.rb
@@ -41,6 +41,7 @@ module Authlogic
           klass.class_eval do
             include InstanceMethods
             validates_uniqueness_of :single_access_token,
+              case_sensitive: true,
               if: :will_save_change_to_single_access_token?
 
             before_validation :reset_single_access_token, if: :reset_single_access_token?